First revision in quite a while! Most of the modifications here have been generic cleanups in the math headers and new entities. I finally got the brush model code written, and the platforms and two doors have already been ported. Next is going to be the func_door_secret, then I'm going to write the trigger code and get the triggers ported over. cc_utils was added, it's basically a g_utils but CleanCode style. Fixed a bug in CRailGunShot where you could hit yourself with a railgun by crouching inside a body and firing. Fixed the bullets hitting the sky. Added CUsableEntity, CBlockableEntity, and fixed up the entity flags which weren't always being added correctly.

+++ source/g_turret.cpp
@@ -173,7 +173,7 @@ void turret_breach_think (edict_t *self)
 
 		// x & y
 		angle = self->state.angles[1] + self->owner->move_origin[1];
-		angle *= (M_PI*2 / 360);
+		angle *= (((float)M_PI)*2 / 360);
 		target[0] = SnapToEights(self->state.origin[0] + cosf(angle) * self->owner->move_origin[0]);
 		target[1] = SnapToEights(self->state.origin[1] + sinf(angle) * self->owner->move_origin[0]);
 		target[2] = self->owner->state.origin[2];
